requirements.txt
streamlit
pypdf
langchain==0.1.20
langchain-community==0.0.38
langchain-text-splitters==0.0.1
sentence-transformers
faiss-cpu
ollama
